Refactor TestKitHelper
