+++ sip-test-kit/src/main/java/de/ikor/sip/foundation/testkit/util/TestKitHelper.java
@@ -2,29 +2,16 @@ package de.ikor.sip.foundation.testkit.util;
 
 import de.ikor.sip.foundation.core.util.SIPExchangeHelper;
 import de.ikor.sip.foundation.testkit.configurationproperties.models.EndpointProperties;
-import de.ikor.sip.foundation.testkit.configurationproperties.models.MessageProperties;
 import de.ikor.sip.foundation.testkit.workflow.givenphase.Mock;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.camel.*;
 import org.apache.camel.builder.ExchangeBuilder;
-import org.apache.camel.support.MessageHelper;
+
+import static org.apache.camel.builder.ExchangeBuilder.anExchange;
 
 /** Utility class that changes the {@link Exchange} */
 @Slf4j
 public class TestKitHelper extends SIPExchangeHelper {
-
-  /**
-   * Creates a {@link MessageProperties} from the {@link Exchange}
-   *
-   * @param exchange that should be mapped
-   * @return serializable message properties
-   */
-  public static MessageProperties mapToMessageProperties(Exchange exchange) {
-    MessageProperties messageProperties = new MessageProperties();
-    messageProperties.setBody(MessageHelper.extractBodyAsString(exchange.getMessage()));
-    messageProperties.setHeaders(filterNonSerializableHeaders(exchange));
-    return messageProperties;
-  }
 
   /**
    * Get route id from the {@link Exchange}
@@ -75,17 +62,6 @@ public class TestKitHelper extends SIPExchangeHelper {
   }
 
   /**
-   * Creates an empty {@link Exchange}
-   *
-   * @param camelContext is necessary for creation
-   * @return exchange
-   */
-  public static Exchange createEmptyExchange(CamelContext camelContext) {
-    ExchangeBuilder exchangeBuilder = ExchangeBuilder.anExchange(camelContext);
-    return exchangeBuilder.build();
-  }
-
-  /**
    * Create exchange from test definition
    *
    * @param properties with route id and payload for exchange body
@@ -95,10 +71,10 @@ public class TestKitHelper extends SIPExchangeHelper {
   public static Exchange parseExchangeProperties(
       EndpointProperties properties, CamelContext camelContext) {
     if (properties == null) {
-      return createEmptyExchange(camelContext);
+      return anExchange(camelContext).build();
     }
     ExchangeBuilder exchangeBuilder =
-        ExchangeBuilder.anExchange(camelContext).withBody(properties.getMessage().getBody());
+        anExchange(camelContext).withBody(properties.getMessage().getBody());
     properties.getMessage().getHeaders().forEach(exchangeBuilder::withHeader);
     exchangeBuilder.withProperty(Mock.ENDPOINT_ID_EXCHANGE_PROPERTY, properties.getEndpoint());
     return exchangeBuilder.build();
