+++ sip-test-kit/src/main/java/de/ikor/sip/foundation/testkit/workflow/givenphase/ReportActivityProxyExtension.java
@@ -1,10 +1,10 @@
 package de.ikor.sip.foundation.testkit.workflow.givenphase;
 
+import static de.ikor.sip.foundation.testkit.configurationproperties.models.MessageProperties.mapToMessageProperties;
 import static javax.xml.bind.DatatypeConverter.parseBoolean;
 
 import de.ikor.sip.foundation.core.proxies.ProcessorProxy;
 import de.ikor.sip.foundation.core.proxies.extension.ProxyExtension;
-import de.ikor.sip.foundation.testkit.util.TestKitHelper;
 import de.ikor.sip.foundation.testkit.workflow.TestCase;
 import de.ikor.sip.foundation.testkit.workflow.TestExecutionStatus;
 import de.ikor.sip.foundation.testkit.workflow.whenphase.routeinvoker.RouteInvoker;
@@ -26,7 +26,7 @@ public class ReportActivityProxyExtension implements ProxyExtension {
     findTestReport(original)
         .getMockReport(proxy.getId())
         .setActual(original)
-        .setActualMessage(TestKitHelper.mapToMessageProperties(original));
+        .setActualMessage(mapToMessageProperties(original));
   }
 
   private String getTestName(Exchange original) {
