+++ sip-test-kit/src/main/java/de/ikor/sip/foundation/testkit/workflow/reporting/model/MockReport.java
@@ -1,6 +1,6 @@
 package de.ikor.sip.foundation.testkit.workflow.reporting.model;
 
-import static de.ikor.sip.foundation.testkit.util.TestKitHelper.mapToMessageProperties;
+import static de.ikor.sip.foundation.testkit.configurationproperties.models.MessageProperties.mapToMessageProperties;
 
 import de.ikor.sip.foundation.testkit.configurationproperties.models.MessageProperties;
 import de.ikor.sip.foundation.testkit.workflow.thenphase.result.ValidationResult;
